fix: update `webpackAssets` structure
This patch update `webpackAssets` structure to include all of chunks,
bundle which is consumed by package.

And unify webpack config.
